scan-manager-flask-app/requirements.txt
Flask
nibabel
numpy
Pillow
requests
gunicorn
flask_login
firebase_admin
 